can start monthly saving

+++ src/app/account/accountSettings/DeleteSettingsController.java
@@ -5,11 +5,18 @@ import app.Entities.CT;
 import app.db.DB;
 import javafx.fxml.FXML;
 import javafx.scene.control.ChoiceBox;
+import javafx.scene.control.Label;
+import javafx.scene.control.TextField;
+import javafx.scene.paint.Color;
 
 public class DeleteSettingsController {
 
     @FXML
     ChoiceBox<Account> accountToDelete;
+    @FXML
+    TextField reenterName;
+    @FXML
+    Label warningMessage;
 
     @FXML
     private void initialize() {
@@ -24,13 +31,19 @@ public class DeleteSettingsController {
     }
 
     @FXML
-    void deleteAccount(){
+    void deleteAccount() {
+        if (!reenterName.getText().equals(accountToDelete.getValue().getName())) {
+            warningMessage.setTextFill(Color.RED);
+            warningMessage.setText("Namnet matchar inte kontot");
+            return;
+        }
+
         DB.deleteAccount(accountToDelete.getValue().getAccountNr());
         CT.navController.loadHome();
     }
 
     @FXML
-    void cancelDelete(){
+    void cancelDelete() {
         CT.navController.loadHome();
     }
